Create Genome chaos
During the progression of cancer cells, the degree of genome instability increases leading to genome chaos in populations of malignant cells. While normally chaos is associated with the ergodicity, i.e., the state when the time average of relevant parameters is equal to phase space aver-age, the situation with cancer propagation is more complex. Chromothripsis, a catastrophic massive genomic rearrangement, is observed in many types of cancer leading to increased mutation rates. We present an entropic model of genome chaos and ergodicity and experimental evidence that in-creasing the degree of chaos beyond the non-ergodic threshold may lead to the self-destruction of the tumor cells. We study time and population averages of chromothripsis frequency in cloned rhabdomyosarcoma of rats stem cells. Clones with frequency above 10% result in cells apoptosis possibly due to mutations in the BCL2 gene. Potentially this can be used for suppressing cancer cells by shifting them into non-ergodic proliferation regime.